refactor(daemon): split runtime snapshot rendering
